node: more secure state file handling
Currently, plugins which run as root mix their state files in the same
directory as non-root plugins. The state directory is owned by munin:munin and
is group-writable. Because of these facts, it is possible for an attacker who
operates as user munin to cause a root-run plugin to run arbitrary code as
root.

This commit introduces :
- a new plugin state directory root, owned by uid 0
- each plugin runs in its own UID plugin state directory, owned by the said UID.

Closes: #1234, D#679897, R#849830, CVE-2012-3512