asterisk: introduce a new, multigraph-capable plugin for asterisk.
The new plugin replaces all seven of the old plugins, with a single,
common and collapsed configuration, by drawing multiple graphs with a
single set of transactions to the server.

It also supports autoconf (which was not supported by the old asterisk
plugins), and wildcard/virtual node mode, when monitoring a remote
server.