squid: introduce new Squid plugin to replace the old ones.
Again this plugin uses multigraph and supports virtualnode, and
collapses the five previous squid plugins, reducing the number of
requests coming in for the proxy itself.

Neither output nor configuration ports directly to the new plugin,
unfortunately, but it does allow for saner output and store (in
particular killing cdefs).