entropy: rewrite plugin in Perl.
This allows us to not have to execute multiple commands to be able to
get the entropy (which could, by themselves, deplete it!).

Additionally, this version fixes the vlabel, and sets the upper limit,
as well as the maximum reading values, to the poolsize.

Since the poolsize is constant for a given kernel, this allows to have
a better, proportioned view of how the entropy varies in the system.