load: rewrite plugin in Perl.
The old load plugin looked more like an example plugin that a real
one, but since it's one of the most commonly used ones, it's a good
idea to have this _not_ fork.

The rewritten plugin outputs all three load values, as sometimes the
trend with them helps identifying problems. It also supports proper
autoconfiguration (when /proc is not accessible), and dirtyconfig.